Files and Temporary Overview Functions
Added: line amount in the start of products.txt that is now loaded and saved in the program.
Added: Loading and saving can now handle if the products.txt file does not already exist.
Temporary: Functions inside overview that can add and remove product entries from the products.txt file
Added: The index of the item in listItem under UI is now dynamic